move a large part of the api demo
