+++ docs/api/kea/githubComponent.js
