+++ package.json
@@ -14,9 +14,10 @@
     "classnames": "^2.3.1",
     "kea": "2.5.1",
     "kea-loaders": "^0.5.1",
+    "kea-router": "^1.0.2",
     "posthog-docusaurus": "^1.0.3",
-    "react": "^16.8.4",
-    "react-dom": "^16.8.4",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-gist": "^1.2.2",
     "react-redux": "^7.2.5",
     "redux": "^4.1.1",
@@ -36,7 +37,7 @@
   },
   "devDependencies": {
     "@docusaurus/module-type-aliases": "^2.0.0-beta.7",
-    "husky": "^6.0.0",
+    "husky": "^7.0.2",
     "kea-typegen": "^1.2.2",
     "prettier": "2.4.1",
     "pretty-quick": "^3.1.1",
